Probably this will be the final commit to this project. After research im gonna use apache airflow to create a pipeline that makes the clean data, feature engineering and model training
